+++ CHANGELOG.md
@@ -3,6 +3,40 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.145.0-next.64](https://github.com/carbon-design-system/carbon-addons-iot-react/compare/v2.145.0-next.63...v2.145.0-next.64) (2021-04-28)
+
+
+### Bug Fixes
+
+* **emptytable:** update the empty table state for filter usecase ([8ea6420](https://github.com/carbon-design-system/carbon-addons-iot-react/commit/8ea64206926a8eff2648cd68fed3d299e5cf5cdb))
+
+
+
+
+
+# [2.145.0-next.63](https://github.com/carbon-design-system/carbon-addons-iot-react/compare/v2.145.0-next.62...v2.145.0-next.63) (2021-04-28)
+
+
+### Features
+
+* **tablecolumn:** add optional tooltip to table column ([f43bb02](https://github.com/carbon-design-system/carbon-addons-iot-react/commit/f43bb02bdd9ffd1529ca9f0013ccd5872636d4f4))
+
+
+
+
+
+# [2.145.0-next.62](https://github.com/carbon-design-system/carbon-addons-iot-react/compare/v2.145.0-next.61...v2.145.0-next.62) (2021-04-28)
+
+
+### Bug Fixes
+
+* **progressindicator:** add title attribute for text ellipsis ([0e6a933](https://github.com/carbon-design-system/carbon-addons-iot-react/commit/0e6a9333d233c4b6f1fdcc19113ede82f3429ae1))
+* **progressindicator:** revert span to p element ([7cec215](https://github.com/carbon-design-system/carbon-addons-iot-react/commit/7cec215aa65c1973cc2b649a4108d05ddd9a84ab))
+
+
+
+
+
 # [2.145.0-next.61](https://github.com/carbon-design-system/carbon-addons-iot-react/compare/v2.145.0-next.60...v2.145.0-next.61) (2021-04-27)